Pull bundle.ts out of config.js and make types.ts
