+++ packages/fastify-vite/src/config.js
@@ -1,13 +1,8 @@
-const {
-  resolve,
-  resolveIfRelative,
-  exists,
-  read,
-} = require('./ioutils.cts')
 const { createClientEnvironment, createSSREnvironment } = require('./config/environments.ts')
 const { DefaultConfig } = require('./config/defaults.ts')
 const { resolveClientModule, resolveRoot } = require('./config/paths.ts')
 const { resolveViteConfig } = require('./config/vite-config.ts')
+const { resolveSSRBundle, resolveSPABundle } = require('./config/bundle.ts')
 
 async function configure(options = {}) {
   const defaultConfig = { ...DefaultConfig }
@@ -44,74 +39,6 @@ async function configure(options = {}) {
     vite.fastify.clientModule ?? config.clientModule ?? resolveClientModule(vite.root)
 
   return config
-}
-
-
-async function resolveSSRBundle({ dev, vite, root }) {
-  const bundle = {}
-  let clientOutDir
-
-  if (!dev) {
-    if (vite.fastify) {
-      clientOutDir = resolveIfRelative(
-        vite.fastify.outDirs.client,
-        await getApplicationRootDir(root),
-      )
-    } else {
-      // Backwards compatibility for projects that do not use the viteFastify plugin.
-      bundle.dir = resolveIfRelative(vite.build.outDir, vite.root)
-      clientOutDir = resolve(bundle.dir, 'client')
-    }
-
-    const indexHtmlPath = resolve(clientOutDir, 'index.html')
-    if (!exists(indexHtmlPath)) {
-      return
-    }
-    bundle.indexHtml = await read(indexHtmlPath, 'utf8')
-    const manifestPaths = [
-      // Vite v4 and v5
-      resolve(clientOutDir, 'ssr-manifest.json'),
-      // Vite v6 Beta
-      resolve(clientOutDir, '.vite/ssr-manifest.json'),
-      // Vite v6
-      resolve(clientOutDir, '.vite/manifest.json'),
-    ]
-    for (const manifestPath of manifestPaths) {
-      if (exists(manifestPath)) {
-        bundle.manifest = require(manifestPath)
-      }
-    }
-  } else {
-    bundle.manifest = []
-  }
-  return bundle
-}
-
-async function resolveSPABundle({ dev, vite, root }) {
-  const bundle = {}
-  if (!dev) {
-    let clientOutDir
-
-    if (vite.fastify) {
-      clientOutDir = resolveIfRelative(
-        vite.fastify.outDirs.client,
-        await getApplicationRootDir(root),
-      )
-    } else {
-      // Backwards compatibility for projects that do not use the viteFastify plugin.
-      bundle.dir = resolveIfRelative(vite.build.outDir, vite.root)
-      clientOutDir = resolve(bundle.dir, 'client')
-    }
-
-    const indexHtmlPath = resolve(clientOutDir, 'index.html')
-    if (!exists(indexHtmlPath)) {
-      return
-    }
-    bundle.indexHtml = await read(indexHtmlPath, 'utf8')
-  } else {
-    bundle.manifest = []
-  }
-  return bundle
 }
 
 module.exports = {
