+++ packages/fastify-vite/src/ioutils.ts
